requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
requests
python-dotenv
tqdm
groq
feedparser
streamlit

